[TF-TRT] `trt_convert.py` split in v1 and v2
